Move settings views into their own file.
