+++ zproject/urls.py
@@ -113,9 +113,9 @@ urlpatterns += patterns('zerver.views',
     url(r'^json/send_message$',             'messages.json_send_message'),
     url(r'^json/invite_users$',             'json_invite_users'),
     url(r'^json/bulk_invite_users$',        'json_bulk_invite_users'),
-    url(r'^json/settings/change$',          'json_change_settings'),
-    url(r'^json/notify_settings/change$',   'json_change_notify_settings'),
-    url(r'^json/ui_settings/change$',       'json_change_ui_settings'),
+    url(r'^json/settings/change$',          'user_settings.json_change_settings'),
+    url(r'^json/notify_settings/change$',   'user_settings.json_change_notify_settings'),
+    url(r'^json/ui_settings/change$',       'user_settings.json_change_ui_settings'),
     url(r'^json/subscriptions/remove$',     'streams.json_remove_subscriptions'),
     url(r'^json/subscriptions/add$',        'streams.json_add_subscriptions'),
     url(r'^json/subscriptions/exists$',     'streams.json_stream_exists'),
@@ -126,7 +126,7 @@ urlpatterns += patterns('zerver.views',
     url(r'^json/get_active_statuses$',      'json_get_active_statuses'),
     url(r'^json/tutorial_send_message$',    'json_tutorial_send_message'),
     url(r'^json/tutorial_status$',          'json_tutorial_status'),
-    url(r'^json/change_enter_sends$',       'json_change_enter_sends'),
+    url(r'^json/change_enter_sends$',       'user_settings.json_change_enter_sends'),
     url(r'^json/get_profile$',              'json_get_profile'),
     url(r'^json/report_error$',             'json_report_error'),
     url(r'^json/report_send_time$',         'json_report_send_time'),
@@ -141,9 +141,9 @@ urlpatterns += patterns('zerver.views',
     url(r'^json/refer_friend$',             'json_refer_friend'),
     url(r'^json/set_alert_words$',          'json_set_alert_words'),
     url(r'^json/set_muted_topics$',         'json_set_muted_topics'),
-    url(r'^json/set_avatar$',               'json_set_avatar'),
-    url(r'^json/time_setting$',             'json_time_setting'),
-    url(r'^json/left_side_userlist$',       'json_left_side_userlist'),
+    url(r'^json/set_avatar$',               'user_settings.json_set_avatar'),
+    url(r'^json/time_setting$',             'user_settings.json_time_setting'),
+    url(r'^json/left_side_userlist$',       'user_settings.json_left_side_userlist'),
 
     # This json format view is used by the LEGACY pre-REST API.  It
     # requires an API key.
@@ -181,8 +181,6 @@ v1_api_and_json_patterns = patterns('zerver.views',
              'POST': 'create_user_backend'}),
     url(r'^users/me$', 'rest_dispatch',
             {'GET': 'get_profile_backend'}),
-    url(r'^users/me/enter-sends$', 'rest_dispatch',
-            {'POST': 'json_change_enter_sends'}),
     url(r'^users/me/pointer$', 'rest_dispatch',
             {'GET': 'get_pointer_backend',
              'PUT': 'update_pointer_backend'}),
@@ -193,8 +191,6 @@ v1_api_and_json_patterns = patterns('zerver.views',
              'DELETE': 'remove_alert_words'}),
     url(r'^realm$', 'rest_dispatch',
             {'PATCH': 'update_realm'}),
-    url(r'^users/me/api_key/regenerate$', 'rest_dispatch',
-            {'POST': 'regenerate_api_key'}),
     url(r'^users/me/presence$', 'rest_dispatch',
             {'POST': 'update_active_status_backend'}),
     # Endpoint used by iOS devices to register their
@@ -234,6 +230,12 @@ v1_api_and_json_patterns = patterns('zerver.views',
             {'GET':  'render_message_backend'}),
     url(r'^messages/flags$', 'rest_dispatch',
             {'POST':  'update_message_flags'}),
+
+) + patterns('zerver.views.user_settings',
+    url(r'^users/me/api_key/regenerate$', 'rest_dispatch',
+        {'POST': 'regenerate_api_key'}),
+    url(r'^users/me/enter-sends$', 'rest_dispatch',
+        {'POST': 'json_change_enter_sends'}),
 
 ) + patterns('zerver.views.streams',
     url(r'^streams$', 'rest_dispatch',
